video_chat_backend/requirements.txt
channels==4.0.0
channels-redis==4.1.0
Django==3.2.12
django-cors-headers==3.12.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.0
